Add Registry model subystem doc; add filter to fix links
Add the documentation of the Registry model subystem.

Add a Pandoc Lua filter to map links ending in ".md" to say ".html" in
generated HTML. Specify the use of the filter in the YAML comments of
all .md files.

Update all links to add the ".md" suffix; these will then get
rewritten by the filter to have the correct ".html" suffix.